SERVER/iot-data-server/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2==2.9.9  # PostgreSQL Python 드라이버 (권장)
psycopg2-binary==2.9.7  # Windows에서 설치 문제 발생 가능
# 대안: Windows에서 psycopg2 설치 문제 시 사용
# pg8000==1.29.8  # 순수 Python PostgreSQL 드라이버
asyncpg==0.28.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Task Scheduler
apscheduler==3.10.4

# Validation and Serialization
pydantic==1.10.13
# pydantic-settings==2.10.1  # Pydantic v2용, v1에서는 불필요
pydantic[email]==1.10.13
email-validator==2.2.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-cov==4.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
